Introduce separate NodeFactory file
